better structure of libs toppingmaker and ilitoppingmaker are now the libs and ProjectTopping and IliProjectTopping the objects (where the second inherits the first to make an ILI thing out of a normal project thing)
