+++ QgisModelBaker/gui/topping_wizard/generation_page.py
@@ -24,14 +24,14 @@ from qgis.PyQt.QtWidgets import QWizardPage
 import QgisModelBaker.utils.gui_utils as gui_utils
 from QgisModelBaker.utils import gui_utils
 
-PAGE_UI = gui_utils.get_ui_class("toppingmaker_wizard/generation.ui")
+PAGE_UI = gui_utils.get_ui_class("topping_wizard/generation.ui")
 
 
 class GenerationPage(QWizardPage, PAGE_UI):
     def __init__(self, parent, title):
         QWizardPage.__init__(self)
 
-        self.toppingmaker_wizard = parent
+        self.topping_wizard = parent
 
         self.setupUi(self)
 
@@ -40,9 +40,7 @@ class GenerationPage(QWizardPage, PAGE_UI):
         self.run_generate_button.clicked.connect(self.generate)
 
     def generate(self):
-        ilidata_file = self.toppingmaker_wizard.topping_maker.bakedycakedy(
-            QgsProject.instance()
-        )
+        ilidata_file = self.topping_wizard.topping.bakedycakedy(QgsProject.instance())
         if ilidata_file:
             self.progress_bar.setValue(100)
             self.progress_bar.setFormat(self.tr("Topping generated 🧁"))
